refactor: simplify sweep runtime and recover publication backlog
