+++ dashboard/wrangler.toml
@@ -91,7 +91,9 @@ EXACT_REVIEW_PUBLICATION_FRESH_LANE_ENABLED = "1"
 EXACT_REVIEW_PUBLICATION_FRESH_LANE_MAX_ITEMS = "2"
 EXACT_REVIEW_PUBLICATION_FRESH_LANE_MAX_AGE_MS = "900000"
 EXACT_REVIEW_PUBLICATION_BATCH_WAIT_MS = "60000"
-EXACT_REVIEW_PUBLICATION_BATCH_DISPATCH_COOLDOWN_MS = "30000"
+# Refill the existing eight bounded publisher slots quickly; a 30s refill
+# serialized batches despite dozens of ready publication records.
+EXACT_REVIEW_PUBLICATION_BATCH_DISPATCH_COOLDOWN_MS = "5000"
 EXACT_REVIEW_PUBLICATION_BATCH_DISPATCH_RESERVATION_MS = "600000"
 REVIEW_OBSERVABILITY_REQUIRED = "0"
 REVIEW_RECOVERY_ENABLED = "0"
